Move network related code to server/net.py
main.py contains now only the code to assemble the application.
